LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/dashed-function-eval-expected.txt


PASS Literal result
PASS Literal result, typed return
FAIL Literal result, typed return, calc assert_equals: expected "13px" but got "calc(12px + 1px)"
FAIL Literal result, typed return, mismatch assert_equals: expected "" but got "12s"
PASS Missing result descriptor
PASS Literal result, empty
PASS result cascading behavior
FAIL Another dashed-function in result assert_equals: expected "12px" but got ""
PASS Unused argument
PASS Single parameter
PASS Multiple parameters
PASS Single parameter, typed
FAIL Typed parameter with calc() assert_equals: expected "101px" but got "calc(100px + 1px)"
PASS Untyped parameter with calc()
FAIL Various typed parameters assert_equals: expected "101px 360deg 1s" but got "calc(100px + 1px) 1turn 1000ms"
PASS Parameter with complex type (auto)
PASS Parameter with complex type (px)
FAIL Passing argument to inner function assert_equals: expected "12px" but got ""
PASS var() in argument resolved before call
FAIL var() in argument resolved before call, typed assert_equals: expected "101px" but got "calc(100px + 1px)"
FAIL Argument captures IACVT due to invalid var() assert_equals: expected "PASS" but got ""
FAIL Argument captures IACVT due to invalid var(), typed assert_equals: expected "PASS" but got ""
FAIL Argument captures IACVT due to type mismatch assert_equals: expected "PASS" but got "red"
PASS Single parameter with default value
PASS Multiple parameters with defaults
PASS Multiple parameters with defaults, typed
FAIL Default referencing another parameter assert_equals: expected "5px 5px" but got "5px var(--x)"
FAIL Default referencing another parameter, local interference assert_equals: expected "17px 5px" but got "17px var(--x)"
FAIL Default referencing another defaulted parameter assert_equals: expected "5px 5px" but got "5px var(--x)"
FAIL Typed default with reference assert_equals: expected "5px 6px" but got ""
FAIL IACVT arguments are defaulted assert_equals: expected "1 2 3" but got ""
FAIL IACVT arguments are defaulted, typed assert_equals: expected "1 2 3" but got ""
FAIL Arguments are defaulted on type mismatch assert_equals: expected "1 2 3" but got "red 2 360deg"
PASS Unused local
PASS Local does not affect outer scope
PASS Substituting local in result
PASS Substituting multiple locals in result
PASS Local referring to another local
PASS Locals appearing after result
PASS Locals cascading behavior
FAIL Custom properties are visible inside function assert_equals: expected "10px" but got ""
FAIL Substitute local from outer scope assert_equals: expected "PASS" but got ""
FAIL Substitute argument from outer scope assert_equals: expected "PASS" but got ""
FAIL Inner argument shadowing outer argument assert_equals: expected "PASS" but got ""
FAIL Inner argument shadowing outer local assert_equals: expected "PASS" but got ""
FAIL Inner local shadowing outer argument assert_equals: expected "PASS" but got ""
FAIL Inner local shadowing outer local assert_equals: expected "PASS" but got ""
FAIL Referencing outer local containing var() assert_equals: expected "1" but got ""
FAIL Referencing outer typed argument assert_equals: expected "10px" but got ""
FAIL Same function with different scopes assert_equals: expected "1 2 3 0" but got ""
FAIL Referencing local two frames up assert_equals: expected "1" but got ""
FAIL IACVT outer local shadows property assert_equals: expected "PASS" but got ""
FAIL Inner function call should see resolved outer locals assert_equals: expected "10px" but got ""
FAIL Inner function call should see resolved outer locals (reverse) assert_equals: expected "10px" but got ""
PASS Parameter shadows custom property
PASS Local shadows parameter
FAIL IACVT argument shadows outer scope assert_equals: expected "PASS" but got ""
FAIL IACVT argument shadows outer scope, typed assert_equals: expected "PASS" but got ""
FAIL IACVT argument shadows outer scope, type mismatch assert_equals: expected "PASS" but got "red"
PASS Missing only argument
PASS Missing one argument of several
FAIL Passing list as only argument assert_equals: expected "1px,2px" but got "{1px,2px}"
FAIL Passing list as first argument assert_equals: expected "1px, 2px | 3px" but got "{1px, 2px} | 3px"
FAIL Passing list as second argument assert_equals: expected "1px | 2px, 3px" but got "1px | {2px, 3px}"
FAIL Passing comma as argument assert_equals: expected "," but got "{,}"
FAIL Passing {} as argument assert_equals: expected "{}" but got "{{}}"
FAIL Passing non-whole-value {} as argument assert_equals: expected "foo{}" but got "{foo{}}"
FAIL Local variable with initial keyword assert_equals: expected "PASS" but got ""
FAIL Local variable with initial keyword, defaulted assert_equals: expected "PASS" but got ""
FAIL Local variable with initial keyword, no value via IACVT-capture assert_equals: expected "PASS" but got ""
FAIL Default with initial keyword assert_equals: expected "PASS" but got ""
FAIL initial appearing via fallback assert_equals: expected "PASS" but got ""
PASS Local variable with inherit keyword
FAIL Local variable with inherit keyword (nested) assert_equals: expected "PASS" but got ""
FAIL Inheriting an invalid value assert_equals: expected "PASS" but got ""
FAIL Default with inherit keyword assert_equals: expected "PASS1 PASS2" but got ""
FAIL Default with inherit keyword (nested) assert_equals: expected "PASS1 PASS2" but got ""
PASS Local with the unset keyword
PASS Local with the revert keyword
PASS Local with the revert-layer keyword
PASS initial keyword left unresolved on result descriptor
FAIL inherit keyword left unresolved on result descriptor assert_equals: expected "PASS" but got "FAIL"
FAIL unset keyword left unresolved on result descriptor assert_equals: expected "PASS" but got "FAIL"
FAIL revert keyword left unresolved on result descriptor assert_equals: expected "PASS" but got "FAIL"
FAIL revert-layer keyword left unresolved on result descriptor assert_equals: expected "PASS" but got "FAIL"
FAIL Keyword can be returned from function into local variable assert_equals: expected "PASS" but got "FAIL2"
PASS Can not return CSS-wide keyword as length
